move rest of the logic to a helper class
