+++ src/MSBuild.Conversion.SDK/MSBuild.Conversion.SDK.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Build.Locator" Version="$(MicrosoftBuildLocatorVersion)" />
+  </ItemGroup>
+
 </Project>
